# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Digital Library API
  version: '1.0'
  description: API for course assignment — AdonisJS implementation. Integrates with Open Library and Gutendex.
servers:
  - url: http://localhost:3333
components:
  securitySchemes:
    ApiToken:
      type: http
      scheme: bearer
      bearerFormat: API_TOKEN
  schemas:
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        year:
          type: integer
        source:
          type: string
        createdBy:
          type: string
      required: [id, title]
security:
  - ApiToken: []
paths:
  /books:
    get:
      summary: List all stored books
      tags: [Books]
      responses:
        '200':
          description: A list of books
    post:
      summary: Create a new book
      tags: [Books]
      security:
        - ApiToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                authors:
                  type: array
                  items:
                    type: string
                year:
                  type: integer
                source:
                  type: string
      responses:
        '201':
          description: Book created
  /books/{id}:
    get:
      summary: Get a book by id
      tags: [Books]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Book found
        '404':
          description: Not found
    put:
      summary: Update a book
      tags: [Books]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                authors:
                  type: array
                  items:
                    type: string
                year:
                  type: integer
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete book (admin required)
      tags: [Books]
      security:
        - ApiToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /external/openlibrary:
    get:
      summary: Search Open Library
      tags: [External]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      security:
        - ApiToken: []
      responses:
        '200':
          description: Search results
  /external/gutendex:
    get:
      summary: Search Gutendex
      tags: [External]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      security:
        - ApiToken: []
      responses:
        '200':
          description: Search results